Tried New Fit Function
I am trying to fit our new lorentzian, where I substituted for Q in terms of its frequency representation. This didn't seem to work though. I did not play with the parameters for very long, but when I tried the new equation on a single file, none of the fits converged. I didn't think this was going to work to begin with, because from my past experimentation, I found that the equation that we are using for Q, didn't seem to representitive of the data. Previously, when I tried using the new lorenzian equation (without making the substitution for Q), I found that I could get the curves to fit well a couple of times, but when I checked the frequency represtation of Q against what my model fit, I found that the frequency representation was off by about a factor of 2.